Reorganized all files into folders and renamed pathways
