+++ src/CSS/Footer.css
@@ -1,6 +1,6 @@
 @font-face {
     font-family: 'Roboto';
-    src: url('./Roboto-Regular.ttf');
+    src: url('../Data/Roboto-Regular.ttf');
     font-weight: normal;
     font-style: normal;
   }
