+++ src/CSS/Home.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: 'Assistant';
-  src: url('./Assistant-Regular.ttf');
+  src: url('../Data/Assistant-Regular.ttf');
   font-weight: normal;
   font-style: normal;
 }
@@ -27,7 +27,7 @@
       ),
       linear-gradient(0deg, rgba(0, 0, 0, 0.05), rgba(0, 0, 0, 0.05)),
       linear-gradient(0deg, rgba(190, 184, 235, 0.1), rgba(190, 184, 235, 0.1)),
-      url(BackgroundImg.jpg);
+      url(../Data/img/BackgroundImg.jpg);
     background-repeat: no-repeat;
     background-size: cover;
     background-attachment: fixed;
