+++ src/CSS/Join-now.css
@@ -1,6 +1,6 @@
 @font-face {
     font-family: 'Assistant';
-    src: url('./Assistant-Regular.ttf');
+    src: url('../Data/Assistant-Regular.ttf');
     font-weight: normal;
     font-style: normal;
   }
@@ -10,7 +10,7 @@
             rgba(190, 184, 235, 0.8)), linear-gradient(0deg,
             rgba(0, 0, 0, 0.05), rgba(0, 0, 0, 0.05)),
         linear-gradient(0deg, rgba(190, 184, 235, 0.1),
-            rgba(190, 184, 235, 0.1)), url(man-on-computer..jpg);
+            rgba(190, 184, 235, 0.1)), url(../Data/img/man-on-computer..jpg);
     background-size: cover;
     background-attachment: fixed;
     display: flex;
