--- src/Join-Now.css
@@ -1,122 +1,0 @@
-@font-face {
-    font-family: 'Assistant';
-    src: url('./Assistant-Regular.ttf');
-    font-weight: normal;
-    font-style: normal;
-  }
-
-#background-image {
-    background: linear-gradient(0deg, rgba(190, 184, 235, 0.8),
-            rgba(190, 184, 235, 0.8)), linear-gradient(0deg,
-            rgba(0, 0, 0, 0.05), rgba(0, 0, 0, 0.05)),
-        linear-gradient(0deg, rgba(190, 184, 235, 0.1),
-            rgba(190, 184, 235, 0.1)), url(man-on-computer..jpg);
-    background-size: cover;
-    background-attachment: fixed;
-    display: flex;
-    justify-content: center;
-}
-
-#js-white-box {
-    display: flex;
-    flex-direction: column;
-    margin: 15%;
-    border: 1.5px solid #3D3061;
-    background-color: #FFFFFF;
-    border-radius: 35px;
-    justify-content: center;
-    align-items: center;
-    width: 400%;
-    height: 200%;
-}
-
-.completion-bar {
-    display: flex;
-    width: 400px;
-}
-
-#center-images {
-    display: flex;
-    justify-content: center;
-}
-
-#content-flex {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.link-boxes {
-    font-family: 'Assistant';
-    border: 2px solid #3D306180;
-    padding-top: 10px;
-    padding-bottom: 10px;
-    padding-left: 20px;
-    padding-right: 20px;
-    box-sizing: border-box;
-    color: #3D3061;
-    font-size: 110%;
-    font-weight: bold;
-    text-decoration: none;
-    border-radius: 20px;
-    background-color: #E5E2ED;
-
-}
-
-.link-boxes:hover {
-    text-decoration: none;
-    color: #E5E2ED;
-    background-color: #3D3061;
-}
-
-#separate-insta-linkedin {
-    margin: 7.5px;
-}
-
-#button-flex {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    margin: 2%;
-    width: 95%;
-}
-
-.click-button-styling {
-    display: flex;
-    border-radius: 12px;
-    border-color: #9C7EF3;
-    background-color: #9C7EF3;
-    padding-left: 2%;
-    padding-right: 2%;
-    padding-top: 1%;
-    padding-bottom: 1%;
-    font-size: 85%;
-    color: #E5E2ED;
-    font-family: 'Assistant';
-    border: 0px;
-}
-
-#join-message {
-    font-family: 'Assistant';
-    color: #3D3061;
-    font-weight: bolder;
-    font-size: 150%;
-    margin: 10px;
-}
-
-#thanks-box {
-    justify-self: center;
-    box-sizing: border-box;
-    border-radius: 20px;
-    font-family: 'Assistant';
-    font-size: 120%;
-    background-color: #BEB8EB80;
-    border: 1.5px solid #5941A9;
-    text-align: center;
-    width: 120%;
-    margin-left: -15px;
-}
-
-#thanks-message {
-    color:#3D3061;
-}
